move useStartAudio
